Consider only on-met actions for conditions
Currently, each `ConditionWatcher` object is associated with two
actions: one executed when the condition mets and another one executed
when the condition stops to met.

In some circumstances, several conditions are needed for the same
device. It could happen that when one of them is not met anymore it
could mean either other one is met or none of them are met. If a not-met
action is associated with each condition, that action could destroy
the work done by on-met handler of another condition.

This change fixes this by removing the not-met action of conditions. In
case of `Bus` class, this feature has been used to model conditions
determining what is the device that is energizing the bus.