Fix a bug in condition evaluation
The tree of conditions was wrongly evaluated when there was more
than one node trying to evaluate the same condition. This may cause
that this node was notified by the event before its siblings, what
means an intermediate, inconsistent state that produces wrong
evaluations.

This change fixes the bug by moving the event notification to the
watcher and making it collect the event status (condition context) and
request the conditions to evaluate it.