Increase the power of ConditionEvaluator
The `ConditionEvaluator` class was only able to associate actions with
boolean conditions. This is a poor feature since most of the times the
conditions should be expressed not in terms of met/not-met but in terms
of “condition `C` mets with evaluation result `E`”, being `E` an input
to the action associated with the watcher. With his approach, it’s
possible to represent more sophisticated conditions as “when BUS `X` is
powered on by supply `Y`, then power this contactor `C` with supply
`Y`”.