Improve condition evaluation even more
Actually `Condition` objects that evaluate on the same type can be
combined together in the same watcher comprising a responsibility chain.
The first one to evaluate triggers its associated action. If none of
them evaluates, a not-met action is triggered. This may be combined
in different ways to have many different ways to check conditions and
react agains them.