Keep supply chain in device state
Instead of storing the device that is supplying the power on bus,
contactor, and TR states, it’s better to store the whole supply chain,
i.e., the sequence of devices that brings the power to the device. That
makes possible to create more sophisticated conditions that check not
only if a element is powered, but the source of that power as well.