Refactor current stuff to avoid contactor simulation
It doesn’t make any sense to model electrical system contactors. They
are not part of the flight crew operation: their existence is not
reflected in the ECAM page of in any other system.

This change eliminates the contactors and refactors the rest of the
module to depend exclusively in generators, buses, TRs and batteries.